Adjust tabby1server to Render MITUS Results
Replace the year breakpoints in tabby1 (2016, 2025, 2050, 2075, 2100)
with breakpoints appropriate for tabby2: 2018, 2020, 2025, 2035, 2050.

Right now the restab2 data imported from MITUS doesn't have
confidence intervals (i.e. no statistic/type column with entries
mean/ci_high/ci_low) or any of the non-base-case scenarios.

The code for renaming the 'statistic' column to 'type' is removed
because the column is named 'type' to begin with when we're importing
the restab2 data and using a temporary patch to avoid having to re-write
the code without confidence intervals:

Instead of re-writing the code without confidence intervals and then
reverting it for later when we have confidence intervals, we have
duplicated the values column as each of mean/ci_low/ci_high. This
is a temporary fix and will be changed later when MITUS produces
these statistics in the restab2 files.