Add More Next/Back Buttons
In the Custom Scenarios Builder, we've added buttons for
advancing the user through the intended workflow. On the first two
TTT interventions buttons are shown to go to the next TTT intervention
("Define Another TTT Intervention"). Similarly on the Program Change
Scenarios and in the Custom Scenarios.

All pages show Go To Outcomes in the Custom Scenarios Builder.

observeEvents are added for each actionButton to call either
updateTabPanel or updateTabsetPanel accordingly.

Many of the actionButtons are given names which end in a numeric
meant to represent their order of appearance in the application
relative to the other actionButtons which serve the same purpose.
This is because if multiple inputs share the same ID on the same page,
it seems that observeEvents stop registering events on those IDs.