Add Interventions to Data Template for Outcomes
CatList is a list containing entries which make up a
template for the data coming out of the reshaper written
in PPML/reshape_mitus_results_for_tabby.

In it, we specify the scenarios listed in the simulation data.
This commit adds intervention_1 through 5 and scenario_1 through 3
to this CatList template.