Setup TTT to use older reshaping code; Fix multiple ttt scenario bug.
A version of the reshaping code was written to work with indices
retrieved at the time of the reshaping call (as opposed to fixed
indices) but we haven't gotten that version working, so MITUS is
adjusting the TTT output to be of the same format as the Program
Changes format.

The ttt scenario bug was that if multiple were run, an error was
generated that said the rows needed to be unique when they were
bound together; that was because the compute_ttt_n reactives
were all being called with n = 1, instead of n = 1, 2, 3.