Move away from using Rmd for changelog/readmore
The Rmarkdown rendering approach that Tabby2 currently uses
was causing problems in production when I tried to deploy it.

Namely, when the rmarkdown::render or knitr::knit step was being
run by the shiny-server, the server found that the shiny user
didn't have permissions to overwrite the
utilities/inst/Rmd/{changelog,readmore2}.md files.

A hacky solution that has kept things running for a few days has
been to manually change the permissions on these .md files to
be writeable by the shiny server, but whenever git updates these
files, it seems that their permissions are over-written,
and thus the permissions problem described re-appears resulting
in a "Error: Could not connect to the server" message being
displayed to the user.