Round & improve the DataTables
In the Estimates / Time Trends / Age Groups
data tables, make sure to round the values to 2 decimal points.

For the estimates, since the plot's scale is transformed,
make sure that the data shown in the Data Table & in the CSV, XLSX
use the actual years, not the ones which are transformed simply
to make the plot look right (introduced in estimatesDataTransformed).